+++ examples/medplum-chart-demo/src/components/DiagnosticReportTask.tsx
@@ -1,0 +1,37 @@
+import { Button, Flex, Modal } from '@mantine/core';
+import { DiagnosticReport, Task } from '@medplum/fhirtypes';
+import { CodeableConceptDisplay, DiagnosticReportDisplay, useMedplum } from '@medplum/react';
+import { IconCircleCheck } from '@tabler/icons-react';
+import React from 'react';
+import { useState } from 'react';
+import { TaskCellProps } from './TaskList';
+
+export function DiagnosticReportModal(props: TaskCellProps) {
+  const [open, setOpen] = useState(false);
+  const [reviewed, setReviewed] = useState(false);
+  const medplum = useMedplum();
+  const report = props.resource as DiagnosticReport;
+  
+  const handleClick = async () => {
+    await medplum.updateResource<Task>({ ...props.task, status: 'completed' });
+    setReviewed(true);
+  };
+  if (reviewed) {
+    return <IconCircleCheck color="#79d290" size={48} />;
+  }
+  return (
+    <>
+      <Button variant="outline" onClick={() => setOpen(true)} sx={{ color: '#4EB180', borderColor: '#4EB180' }}>
+        Review {<CodeableConceptDisplay value={report.code} />}
+      </Button>
+      <Modal opened={open} onClose={() => setOpen(false)} size="xl">
+        <DiagnosticReportDisplay value={report} />
+        <Flex justify={'flex-end'}>
+          <Button mt={8} onClick={handleClick}>
+            Ok
+          </Button>
+        </Flex>
+      </Modal>
+    </>
+  );
+}
